WIP - healthz has an internal server error
