+++ my-app/src/app.ts
@@ -1,7 +1,9 @@
 import { Hono } from "hono";
 import {
+  getCurrentDbRevision,
   type DBClient
 } from "./infrastructure/database.js";
+import { healthz } from "./usecases/healthz.usecase.js";
 
 function getApp(db: DBClient) {
   const app = new Hono();
@@ -14,8 +16,12 @@ function getApp(db: DBClient) {
     return c.json(accounts);
   });
   app.get("/healthz", async (c) => {
-    const healthzPayload = {}; // healthz(await getCurrentDbRevision(db));
-    return c.json(healthzPayload);
+    try {
+      const healthzPayload = healthz(await getCurrentDbRevision(db));
+      return c.json(healthzPayload);
+    } catch (e) {
+      return c.json({error: e}, 500);
+    }
   });
 
   return app;
